added links to the checklists' boxes
